requirements.txt
PyYAML==5.4.1
selenium==3.141.0
urllib3==1.26.6
pandas=1.4.1
beautifulsoup4==4.10.0
lxml==4.8.0